3 - fixing changes requested
