+++ src/index.jsx
@@ -1,0 +1,7 @@
+import App from './App';
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+module.hot && module.hot.accept();
+
+ReactDOM.render(<App />, document.getElementById('main'));
